Refactor extract components from __init__ file
When components that are defined in the __init__ file of a package
are used by other components in the same package, those cannot
be exposed in the __init__ file without causing a dependency loop.

Ideally, the __init__ file would only define the package API, while
leaving its components definition to internal modules.

Notes:

- The SvgLabel class contains some obvious mistakes (e.g. type
  hint in the constructor), but I'm not fixing them now. Instead,
  I'm opening follow up issues.

- The test file doesn't currently ensure full coverage of the module.
  Some other test files complete the missing coverage. This is not
  necessarily an issue, but it is certainly not ideal.

- There are two mistakely identical test cases, that I'll replace
  in a later PR.
